Create README.md
The purpose of this project is to demonstrate the collection, working with, and cleaning of a data set. The goal is to prepare tidy data that can be used for later analysis. The script does the following:

Downloads the dataset if it does not exist yet in the working directory. Loads both the training and test datasets, keeping only those columns which reflect a mean or standard deviation Loads the activity and subject data for each dataset, and merges those columns with the dataset Merges the two datasets Converts the activity and subject columns into factors Creates a tidy dataset that consists of the average (mean) value of each variable for each subject and activity pair.

The end result is shown in the file tidy.txt.